+++ coordinator/src/routes.rs
@@ -19,13 +19,11 @@ use bitcoin::secp256k1::PublicKey;
 use diesel::r2d2::ConnectionManager;
 use diesel::r2d2::Pool;
 use diesel::PgConnection;
-use dlc_manager::contract::contract_input::ContractInput;
 use serde::Deserialize;
 use serde::Serialize;
 use serde_json::json;
 use std::sync::Arc;
 use tokio::sync::broadcast;
-use trade::TradeParams;
 
 pub struct AppState {
     pub node: Node,
@@ -54,7 +52,6 @@ pub fn router(node: Node, pool: Pool<ConnectionManager<PgConnection>>) -> Router
             get(get_order).put(put_order).delete(delete_order),
         )
         .route("/api/orderbook/websocket", get(websocket_handler))
-        .route("/api/trade", post(post_trade))
         .with_state(app_state)
 }
 
@@ -121,19 +118,6 @@ pub async fn get_invoice(State(state): State<Arc<AppState>>) -> Result<Json<Stri
         })?;
 
     Ok(Json(invoice.to_string()))
-}
-
-// TODO: We might want to have our own ContractInput type here so we can potentially map fields if
-// the library changes?
-pub async fn post_trade(
-    State(state): State<Arc<AppState>>,
-    trade_params: Json<TradeParams>,
-) -> Result<Json<ContractInput>, AppError> {
-    let contract_input = state.node.trade(trade_params.0).map_err(|e| {
-        AppError::InternalServerError(format!("Failed to accept trade request: {e:#}"))
-    })?;
-
-    Ok(Json(contract_input))
 }
 
 /// Our app's top level error type.
